update token-info after remove session functionality - try-2

+++ src/controllers/devicesController.ts
@@ -14,10 +14,10 @@ export const getDevicesController = async (req: Request<any, any, any, any>, res
     if (!validateToken) {
         return next(ApiError.UnauthorizedError())
     }
-    // const isTokenExists = await tokensRepository.findTokenByRefreshToken(token)
-    // if (!isTokenExists || isTokenExists.blackList) {
-    //     return next(ApiError.UnauthorizedError())
-    // }
+    const isTokenExists = await tokensRepository.findTokenByRefreshToken(token)
+    if (!isTokenExists || isTokenExists.blackList) {
+        return next(ApiError.UnauthorizedError())
+    }
     const devices = await deviceCollection.find().toArray()
     const deviceMap = (device: WithId<IDevice>) => ({
         deviceId: device.deviceId,
@@ -49,10 +49,10 @@ export const deleteDeviceByIdController = async (req: Request, res: Response, ne
         if (!validateToken) {
             return next(ApiError.UnauthorizedError())
         }
-        // const isTokenExists = await tokensRepository.findTokenByRefreshToken(token)
-        // if (!isTokenExists || isTokenExists.blackList) {
-        //     return next(ApiError.UnauthorizedError())
-        // }
+        const isTokenExists = await tokensRepository.findTokenByRefreshToken(token)
+        if (!isTokenExists || isTokenExists.blackList) {
+            return next(ApiError.UnauthorizedError())
+        }
 
         const removedToken = await tokenCollection.findOne({deviceId: req.params.id})
         if (validateToken._id !== removedToken?.userId) {
@@ -77,10 +77,10 @@ export const deleteAllDevicesExceptCurrentController = async (req: Request, res:
         if (!validateToken) {
             return next(ApiError.UnauthorizedError())
         }
-        // const isTokenExists = await tokensRepository.findTokenByRefreshToken(token)
-        // if (!isTokenExists || isTokenExists.blackList) {
-        //     return next(ApiError.UnauthorizedError())
-        // }
+        const isTokenExists = await tokensRepository.findTokenByRefreshToken(token)
+        if (!isTokenExists || isTokenExists.blackList) {
+            return next(ApiError.UnauthorizedError())
+        }
         await deviceCollection.deleteMany({deviceId: {$ne: validateToken.deviceId}})
         // await tokenCollection.updateMany({refreshToken: token}, {$set: {blackList: true}})
         await tokenCollection.updateMany({refreshToken: token, deviceId: {$ne: validateToken.deviceId}}, {$set: {blackList: true}})
